requirements.txt
numpy==1.20.2
Pillow==8.2.0
scipy==1.6.2